added delete option

+++ app/src/main/java/chromahub/rhythm/app/features/local/presentation/screens/PlayerScreen.kt
@@ -2450,7 +2450,7 @@ fun PlayerScreen(
                                     HapticFeedbackType.LongPress
                                 )
                                 onSeek(
-                                    (currentTimeMs + 10000).coerceAtMost(totalTimeMs.toLong())
+                                    (currentTimeMs + 10000).coerceAtMost(totalTimeMs)
                                         .toFloat() / totalTimeMs
                                 )
                             },
